+++ src/components/cars/orange-car/constants.js
@@ -4,8 +4,9 @@ import car from '../assets/car.png';
 import explosion from '../assets/explosion.png';
 
 const PARAMS = {
+  INITIAL_COORDINATES: [6000, 6450],
   IMAGES: {
-    ORANGE_CAR: {
+    PLAYER_CAR: {
       id: nanoid(),
       img: car,
       frameSize: { frameWidth: 256, frameHeight: 256 },
@@ -21,4 +22,13 @@ const PARAMS = {
   },
 };
 
-export default PARAMS;
+const controlKeys = {
+  up: 'W',
+  down: 'S',
+  left: 'A',
+  rigth: 'D',
+  stop: 'SPACE',
+  action: 'ENTER',
+};
+
+export { PARAMS, controlKeys };