+++ src/components/map/constants.js
@@ -21,10 +21,58 @@ const PARAMS = {
     frameWidth: 64,
     frameHeight: 64,
   },
-  frameBorders: {
+  frameWater: {
     start: 43,
     end: 45,
   },
+
+  nameBorder: 'borderCameras',
+  positionBorder: {
+    x: 1145,
+    y: 5,
+  },
+  sizeBorder: {
+    width: 210,
+    height: 210,
+  },
+  backgroundBorder: {
+    r: 255,
+    g: 165,
+    b: 0,
+    a: 1,
+  },
+  fadeTimeBorder: 500,
+  alphaBorder: 1,
+
+  nameMinimap: 'miniCameras',
+  positionMinimap: {
+    x: 1150,
+    y: 10,
+  },
+  sizeMinimap: {
+    width: 200,
+    height: 200,
+  },
+  backgroundMinimap: {
+    r: 139,
+    g: 69,
+    b: 19,
+    a: 1,
+  },
+  tintMinimap: {
+    topLeft: 0xf0e68c,
+    topRight: 0xf0e68c,
+    bottomLeft: 0xf0e68c,
+    bottomRight: 0xf0e68c,
+  },
+  fadeTimeMinimap: 500,
+  alphaMinimap: 0.3,
+
+  minimapZoom: 0.09,
+  scrollXMinimap: 8000,
+  scrollYMinimap: 8000,
+  transparentMinimap: false,
+  tintFillMinimap: true,
 };
 
 export default PARAMS;