+++ src/scenes/constants.js
@@ -4,6 +4,15 @@ import logo from './assets/images/logo.png';
 import cover from './assets/images/cover.jpg';
 import stats from './assets/images/statsbg.jpg';
 import music from './assets/sounds/title.mp3';
+import effect from './assets/sounds/menu.mp3';
+import start from './assets/sounds/RespectIs.mp3';
+import city from './assets/sounds/goroda.mp3';
+
+import controlSprite from '../components/screen-controller/assets/control.png';
+import directionThumpSprite from '../components/screen-controller/assets/directionThump.png';
+import shootThumpSprite from '../components/screen-controller/assets/shootThump.png';
+import actionSprite from '../components/screen-controller/assets/action.png';
+import runSprite from '../components/screen-controller/assets/run.png';
 
 const PARAMS = {
   SCENES: {
@@ -15,12 +24,13 @@ const PARAMS = {
     buttonsScene: nanoid(),
     timeQuestScene: nanoid(),
     score: nanoid(),
+    instructionScene: nanoid(),
   },
   IMAGES: {
     LOGO: {
       id: nanoid(),
       img: logo,
-      coord: [680, 200],
+      coord: [680, 160],
       size: [400, 500],
     },
     COVER: {
@@ -40,6 +50,18 @@ const PARAMS = {
     MENU: {
       id: nanoid(),
       file: music,
+    },
+    EFFECT: {
+      id: nanoid(),
+      file: effect,
+    },
+    START: {
+      id: nanoid(),
+      file: start,
+    },
+    CITY: {
+      id: nanoid(),
+      file: city,
     },
   },
   LINES: {
@@ -68,7 +90,7 @@ const PARAMS = {
       green: '#004d00',
       yellow: '#ffa500',
     },
-    coord: [660, 400, 470, 540, 600, 700, 820],
+    coord: [660, 350, 420, 520, 570, 700, 820, 470, 620],
     text: {
       start: 'START GAME',
       fullscreen: 'FULLSCREEN:',
@@ -76,10 +98,17 @@ const PARAMS = {
       on: 'on',
       score: 'SCORE',
       login: 'SIGN IN',
+      sound: 'MUSIC:',
+
+      instruction: 'CONTROLS',
     },
     switchStyle: {
       font: '60px gta',
       fill: '#ff0000',
+    },
+    switchMusicStyle: {
+      font: '60px gta',
+      fill: '#004d00',
     },
     loginStyle: {
       font: '90px gta',
@@ -93,12 +122,40 @@ const PARAMS = {
       font: '60px gta',
       fill: '#ffa500',
     },
+    textControlsStyle: {
+      font: '44px gta',
+      fill: '#ffffff',
+    },
     shadow: [3, 1, '#000000'],
   },
   CAMERA: {
     fadeTime: 650,
   },
   originCenter: [0.5, 0.5],
+
+  spritesJoystick: {
+    controlId: nanoid(),
+    controlPath: controlSprite,
+
+    directionId: nanoid(),
+    directionPath: directionThumpSprite,
+
+    shootId: nanoid(),
+    shootPath: shootThumpSprite,
+
+    actionId: nanoid(),
+    actionPath: actionSprite,
+
+    runId: nanoid(),
+    runPath: runSprite,
+  },
+
+  textJoystick: {
+    firstJoystick: 'control character movement',
+    secondJoystick: 'line of fire',
+    actionButton: 'action',
+    runButton: 'run',
+  },
 };
 
 export default PARAMS;